+++ src/ArrDash/Services/ServiceCredentialsPreviewService.cs
@@ -36,6 +36,7 @@ public sealed class ServiceCredentialsPreviewService
         ApplyIfSet(_overlay.SlskdApiKey, v => options.Slskd.ApiKey = v);
         ApplyIfSet(_overlay.PlexToken, v => options.Plex.Token = v);
         ApplyIfSet(_overlay.EmbyApiKey, v => options.Emby.ApiKey = v);
+        ApplyIfSet(_overlay.JellyfinApiKey, v => options.Jellyfin.ApiKey = v);
         ApplyIfSet(_overlay.SonarrUrl, v => options.Sonarr.Url = v);
         ApplyIfSet(_overlay.RadarrUrl, v => options.Radarr.Url = v);
         ApplyIfSet(_overlay.LidarrUrl, v => options.Lidarr.Url = v);
@@ -44,6 +45,7 @@ public sealed class ServiceCredentialsPreviewService
         ApplyIfSet(_overlay.SlskdUrl, v => options.Slskd.Url = v);
         ApplyIfSet(_overlay.PlexUrl, v => options.Plex.Url = v);
         ApplyIfSet(_overlay.EmbyUrl, v => options.Emby.Url = v);
+        ApplyIfSet(_overlay.JellyfinUrl, v => options.Jellyfin.Url = v);
     }
 
     private static void ApplyIfSet(string? value, Action<string> apply)